fix: move RequestMethods from types to constants
