+++ src/types.ts
@@ -1,9 +1,3 @@
-export enum RequestMethods {
-  get = 'GET',
-  post = 'POST',
-  put = 'PUT',
-  delete = 'DELETE',
-}
 export type EndpointInfo = {
   full: string
   short: string
